Implement the parser (#1)
* Add initial parser implementation

For now it is a Recursive Descent Parser with a simple Panic Mode
error recovery system. Let's see how it works.

Also add new line at the end of Cargo.toml

* Make parser::definition() report all errors.

Before the function consumed the tokens and supposed they were what
we wanted. Now it checks them and returns an error if an incorrect
token was found.

But is an ugly as hell function. Try to make it nicer.

* Change test to check ::parameters return value.

* Remove trailing spaces

* Add comment and a bunch of tests.

Comment the ::parser::definition function and write many tests.

* Finish first implementation of the parser.

* Transform Parser into an Iterator.